refactor(tab): file names to follow convention
