+++ src/components/Tab/sgds-tab-group.scss
