+++ src/components/Tab/sgds-tab-panel.scss
